The reason I decided to desist from adding thousands of instances of a some Interpreter actor to perform vesper's operations and use a combination of a VesperInterpreter, detatch() directive, and the onComplete Future directive is the following:
`eval` might take some time come up with its Future result (e.g. because it needs to parse the Java code in order to perform the desired refactoring). It then returns the Future because it will again take some time to actually create a `ChangeSummary`.  I.e. we are potentially dealing with two separate "costly" operations here:
1. Parsing code and then performing refactoring
2. Creating a ChangeSummary object

The first operation is "protected" with the detach() and the second one by returning a Future.